Install LbireOffice via yum repository
The previous incarnation grabbed the RPMs from a tarball, and called
yum to install them. When re-running the role, yum would declare
there is nothing to do and error. While this sounds like a bug in
yum, it caused the role to fail.

Using a repository, and attempting to reinstall via yum results in
yum declaring there is nothing to do and exiting with success.

The previous incarnation installed _all_ the RPMs, and this one does as
well. It uses a yum trick to list all the RPMs in a given repository
to blanket install all of them.